@@ -1,5 +1,6 @@
 # coding: us-ascii
+# frozen_string_literal: true
 
 module Declare
-  VERSION = '0.1.0'.freeze
+  VERSION = '0.2.0'
 end
