@@ -1,8 +1,10 @@
 # coding: us-ascii
-require 'test/unit'
+# frozen_string_literal: true
+
+require_relative 'helper'
 
 class Test_Decalre_Behavior_Mixed < Test::Unit::TestCase
-  CODE = <<"EOD"
+  CODE = <<"RUBY"
 class Person
   attr_reader :name, :birth
 
@@ -35,12 +37,12 @@ The Person.new('John') do |john|
     kind_of Time
   end
 end
-EOD
+RUBY
 
   def test_result
-    out = `ruby -w 2>&1 <<EOD
+    out = `ruby -w 2>&1 <<RUBY_CODE
 #{CODE}
-EOD`
+RUBY_CODE`
     assert_equal(2, $?.exitstatus)
     assert_match(/3 scopes, 11 assertions, 2 failures/, out)
   end
